consolidate tradermade code into a single endpoint
