+++ packages/sources/tradermade/src/adapter.ts
@@ -1,30 +1,43 @@
-import { Builder } from '@chainlink/ea-bootstrap'
 import {
   Config,
   ExecuteWithConfig,
   MakeWSHandler,
   AdapterRequest,
-  APIEndpoint,
   ExecuteFactory,
 } from '@chainlink/types'
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { makeConfig, DEFAULT_WS_API_ENDPOINT } from './config'
-import * as endpoints from './endpoint'
-
-export const execute: ExecuteWithConfig<Config> = async (request, context, config) => {
-  return Builder.buildSelector(request, context, config, endpoints)
-}
-
-export const endpointSelector = (request: AdapterRequest): APIEndpoint =>
-  Builder.selectEndpoint(request, makeConfig(), endpoints)
-
-export const makeExecute: ExecuteFactory<Config> = (config) => {
-  return async (request, context) => execute(request, context, config || makeConfig())
-}
+import { makeConfig, DEFAULT_WS_API_ENDPOINT, NAME } from './config'
 
 const customParams = {
   base: ['base', 'from', 'symbol', 'market'],
   quote: ['quote', 'to', 'market', 'convert'],
+}
+
+export const execute: ExecuteWithConfig<Config> = async (input, _, config) => {
+  const validator = new Validator(input, customParams)
+  if (validator.error) throw validator.error
+
+  Requester.logConfig(config)
+
+  const jobRunID = validator.validated.id
+  const symbol = (validator.overrideSymbol(NAME) as string).toUpperCase()
+  const quote = (validator.validated.data.quote || '').toUpperCase()
+  const currency = `${symbol}${quote}`
+
+  const params = {
+    ...config.api.params,
+    currency,
+  }
+
+  const options = { ...config.api, params }
+
+  const response = await Requester.request(options)
+  response.data.result = Requester.validateResultNumber(response.data, ['quotes', 0, 'mid'])
+  return Requester.success(jobRunID, response)
+}
+
+export const makeExecute: ExecuteFactory<Config> = (config) => {
+  return async (request, context) => execute(request, context, config || makeConfig())
 }
 
 export const makeWSHandler = (config?: Config): MakeWSHandler => {
@@ -50,7 +63,6 @@ export const makeWSHandler = (config?: Config): MakeWSHandler => {
       connection: {
         url: defaultConfig.api.baseWsURL || DEFAULT_WS_API_ENDPOINT,
       },
-      shouldNotServeInputUsingWS: (input: AdapterRequest) => input.data.endpoint !== 'forex',
       subscribe: (input: AdapterRequest) => getSubscription(getPair(input)),
       unsubscribe: () => null, // Tradermade does not support unsubscribing.
       subsFromMessage: (message) => {
